refactor: fix test failures and remove dyn dispatching

+++ src/parser/nfsv3/read.rs
@@ -3,12 +3,12 @@
 use std::io::Read;
 
 use crate::parser::nfsv3::file;
-use crate::parser::primitive::u64;
+use crate::parser::primitive::{u32, u64};
 use crate::parser::Result;
 use crate::vfs::read;
 
 pub fn args(src: &mut impl Read) -> Result<read::Args> {
-    Ok(read::Args { file: file::handle(src)?, offset: u64(src)?, count: u64(src)? })
+    Ok(read::Args { file: file::handle(src)?, offset: u64(src)?, count: u32(src)? })
 }
 
 #[cfg(test)]
